Revert "Add rendering component function (#60)"
This reverts commit c94be07c477c5074d29dfdcc8b5c1f85dea28c07.
