--- jest-setup.js
@@ -1,2 +1,0 @@
-// In your own jest-setup.js (or any other name)
-import '@testing-library/jest-dom';
