--- jest.config.js
@@ -1,8 +1,0 @@
-// Sync object
-/** @type {import('@jest/types').Config.InitialOptions} */
-const config = {
-  verbose: true,
-  setupFilesAfterEnv: ['<rootDir>/jest-setup.js'],
-};
-
-module.exports = config;
