+++ src/components/TestButton.js
@@ -1,0 +1,9 @@
+const TestButton = () => {
+  const testButton = document.createElement('button');
+  testButton.classList.add('testButton');
+  testButton.innerText = "test button"
+
+  return testButton;
+};
+
+export default TestButton;
